backend/requirements.txt
Django~=3.2
djangorestframework~=3.14
djoser~=2.1.0
psycopg2-binary~=2.9
gunicorn~=21.2
python-dotenv~=1.0
django-filter~=23.1
 Pillow~=10.1 